HEYO, big, and maybe last commit
This was a big experiment to learn about apps the exist on both the server and client. I used Marko from Ebay. The biggest issue is lack of easy integration/documentation for working with other libraries. Not having as many eyes on the project is rough.

I think I still like the general approach of an app that is rendered on the server, but parts can be re-rendered on the client easily. I think NextJS is a better option. React for rendering, but provides a framework to wrap a lot of the server/client interaction that I was about to have to implement here.